requirements.txt
python-dotenv
langchain
langchain-google-genai
streamlit
Pillow
Flask